Fix some errors from the rebase.
